✨(lms_handler) an util to route LMS requests to the right lms backend
To support several LMS connections, we have to implement interfaces able to
communicate with each LMS API. So we create a LMSHandler class able to select
the right LMS Backend to use according to the course run's resource_link
provided. For now, only an OpenEdX backend has been implemented and only
enrollment management is supported.