🔧(all) run development server on port 8071
Previously, Joanie ran on port 8060 but this port is already used by Marsha.
To be allow to run those services together, we now start Joanie development
server on port 8071. Why 8071? Because Richie runs on port 8070 and as Joanie is
strongly linked to Richie why not use a port number near to Richie port.