⬆️(Dockerfile) pin base Docker image to python8-slim-bullseye
Our CI runs on python8-bullseye but our image was built on python8-slim
which is a floating version that will always be built on the last version
of Debian available. This can lead to discrepancies between our local
environments and the CI (e.g. incompatible version of libpango recently).
We should pin the slim image to bullseye to avoid these situations.