✨(account) ensure that an user has a main address
Even if user does not mention that address should be main, if this is its first
address, it set as main by default. Furthermore, if user tries to demote its
main address, a ValidationError will be raised. By covering these two use
cases, we ensure that an user has always a main address.