🚨(lms_handler) declare backends folder as package
`lms_handler/backends` was not declared as a package and it can produce some
conflicts if other files have the same name than a file contains in this folder.
So we add `__init__.py` to declare this folder as a package then we fixed
linter issues related to this issue.