🔧(settings) allow configuring media storage backend for production
In production we want to use object storage to store media files.
As a consequence, we can not serve media files publically anymore
and should make the media bucket private. As a consequence images
stored in the media bucket can now only be passed to templates via
the context in the form of base64 encoded strings.