🗃️(core) link enrollments to orders through a many to many relation
Previously, an enrollment could be linked to only one order but it led to issue
when an user had several orders (e.g one valid and another canceled). So from
now, enrollment - order relation is a many to many. In this way, an user can
owns several order relying on the same enrollment.