🗃️(models) remove one-to-many relation between enrollments and order
Previously an enrollment could be linked to only one order. It caused some
conflict in the case where the user enrolled to a course run then the order was
canceled.
To address this issue, we decided to remove relation between an order and an
enrollment as it adds useless complexity. The counterpart of that is a little
increase of db queries count. But we will be able to address this topic later by
creating database indexes.