✨(badges) add Open Badge Factory provider
Open Badge Factory is a popular Open Badge provider that we choose to
implement first while we are not able to generate our own badges.

Note that since the API access token validity is limited (10 hours),
we've implemented an automatic renewal mecanism in case we must generate
badges from the web server and not only from a batch job.

OBF API documentation is available at:
https://openbadgefactory.com/static/doc/obf-api-v1.pdf