🔨(tests) generate a token from a user object
In a future feat, we will be sending emails to users. Writing tests,
we want to make sure that the token used to generate the request
has the same data as the user referring to. In some tests, the user
is created before the generation of the token, it's safer for reliable
tests to create a token based on the data of the user.