✨(backend) update and set email for user
So far the token sent in the email was not updated or even
set up in the first place. In order to save this data, we've
centralized a method `update_or_create` from the user's
model used in the different API calls.
This feat will enable us to send email in a future feature and
save the right email in the payment model.