🚨(tests) fix pylint payment tests
The missing file `__init__.py`in the test folder of the payment module
blocked the check for pylint.
Adding it,  different errors appeared that needed to be fixed.
To test the class BasePaymentBackend, as it's not a class that is
supposed to be instantiated directly, a class inheriting it has been
created in the test class.