♻️(serializer) delete orders from CourseSerializer
CourseSerializer nested orders to give direct access to
these data from the API call. This avoided the front to
do multiple distinct API calls. Nesting data is convenient
until it's not. To maintain a robust API we have decided to
keep the relations simples as much as possible. Orders will
now need to be called from another endpoint.