✨(api) add filtering fields on Course api requests
A course can embed further data. As it implies adding queries,
we don't want them to be executed each time an API call on
`courses` is made. We add a filtering system that will allow the
front to ask for specific fields. Depending on how the API call
is done, results from the serializer will differ.