🚨(badges) split test to fix pylint error that can't be disabled
We was getting error C0103: Attribute name "id" doesn't conform to
snake_case naming style (invalid-name). The disable statement
above the incriminated code line was not taken into account. I
noticed that this test was in fact several tests in one so a split
it which solves the error because we don't need to assign the "id"
attribute.

There is probably a refactoring to do or some declarations to allow
using the protected word "id" as an attribute but this is out of the
scope of the present work.