♻️(organizations) link organizations to courses, products and orders
The many to many relation on the course is to link organizations that have
created a course.
The many to many relation on the product is to link organizations that are
selling the course.
The foreign key on order to link an organization is to identify from which
organization the student buys a course among the organizations linked to
this course.