🚨(core) fix several warnings in test suites
Fix several warning messages displayed in test logs.

In product model test suite, a DateTimeField received a naive datetime while
time zone support is active so a RuntimeWarning was displayed in logs. The
`tzinfo` parameter was missing into datetime.

The `method_whitelist` argument of `urllib3.util.Retry` is deprecated in favor
of `allowed_methods`

The EnrollmentViewSet is paginated but no ordering attribute is provided that
may yield to inconsistent results order.