♻️(payment) rename ProformaInvoice model
In a previous commit we renamed the model from Invoice
to ProformaInvoice thinking invoices from Joanie from
an accounting point of view were not real ones. As
invoices are only generated after payment, they cannot
be a proformat as this kind of document is not related
to payment. So to simplify, we revert and call our
model back Invoice.