✨(orders) create order via the API without specifying the organization
If a product is linked to one and only one organization, we should allow
creating an order via the API without specifying the organization. In
this case, the order is automatically linked to the order.
This will be usefull for retro-compatibility with the existing frontend
in which we don't have to choose an organization when placing an order.