🗃️(enrollment) add BooleanField has_been_created_with_order
Some course run can be available for free and in the same time included in a
product. In case the user has first freely enroll to the course run then
purchase a product including it, we need to know if the user is initially
enrolled freely. So on enrollment creation, client can provide an order id. If
one is provided, the flag `has_been_created_with_order` will be set to True.