✨(back) add a fsm to handle state of Order model
The state of an order was computed. This required extra queries
when we wanted to filter orders. To take into account all the
possible states of an order, code will soon evolve and computing
it each time will add complexity. For these two reasons, we
decided to persist a state in the database using a FSM.
FSM helps simplify code, avoid numerous if-else conditions
by defining states and the connections between them.