🐛(settings) use custom JSONValue class to set COURSE_WEB_HOOKS
Currently, we defined through environment variables the COURSE_WEB_HOOK_URL,
COURSE_WEB_HOOK_SECRET and COURSE_WEB_HOOK_VERIFY then if COURSE_WEB_HOOK_URL
was defined, we tried to set COURSE_WEB_HOOKS settings. But unfortunately this
was not work as when this code was interpreted, COURSE_WEB_HOOK_URL was always
none. In order to fix that, we create a custom JSONValue class which takes a
JSON string as raw value then returns the python representation.