✅(core) fix flaky test models enrollment
`test_models_enrollment_forbid_for_non_listed_course_out_of_scope_of_order` test
was flaky according as by default EnrollmentFactory is_active flag value is
random. But in fact, the ValidationError we want to test is only raised when a
user tries to create an active enrollment. So to fix that, we have to enforce
the `is_active` to True.