♻️(models) move selling organizations to the course/product ManyToMany
The organizations that are selling a product were defined through a
ManyToMany relation on the Product model. This worked for products that
are sold by the same organizations, whatever the course on which they
are sold.

The problem is that we also have products, like for example certificates,
that are sold on many courses but by a different organization on each
course. For this reason, we need to move selling organizations to the
relation model between the Course and the Product.