🐛(core) fix a bug about auto enrollment on order validation
When an order is validated, if the related product has a target course with a
single course run, the user is automatically enrolled to this one. In the case
where there is an existing inactive enrolled for the couple user / course run,
the statement to retrieve enrollment contains a syntax error.