🐛(core) prevent to enroll on several course runs for the same course
We do not want a user to be able to enroll on several opened course runs of the
same course. Currently, we only check if the user is already enrolled on
enrollment creation. One case is not currently covered. Indeed if the user
enrolls to a first course run then unenrolls from it, it will be able to
enroll a second course run and then re-enroll in the first one.