✨(frontend) add CRUD entities pages (5) (#273)
* 🔧(frontend) update gitignore file

Some rules in the file were duplicated. Remove double values

* 🌐(frontend) add react-intl stack

Creation of a component to manage the initialization of language
and change of languages. Adding frontend and backend translation files.

* ✨(frontend) set up the test and mock stack

Add and init jest and MSW (to mock API calls in tests + components)

* 🍱(frontend) add fun logo and avatar example

We display the fun logo in the layout in the left navigation bar

* 💄(frontend) add global variables stack

We need to add global variables to the application like the navbar width

* 🔧(backend) add .env.example

To help a future contributor, we add this file to give
examples of environment variables

* ✨(frontend) create a layout for the Joanie BO

Starting from a blank layout, you have to add a responsive dashboard
type layout

* ✨(frontend) create dashboard layout page component

Added a component to standardize pages in the layout and add sample pages

* 🐛(core) fix duplicate values during a search

When searching, the result contains duplicate values. So we add a distinct

* ➕(frontend) add and update dependencies

Update dependencies for the next.js project

* 🔧(frontend) update jest configuration

The tests were no longer working because of some libs. So we update the jest
configuration

* ✨(frontend) create form components

Using react-hook-form we need to create form components which connects
with RHF using MUI components

* ✨(frontend) create presentational components

We also need to add presentation components to make it easier to use certain
MUI components without repeating themselves.

* 🏷️(frontend) add api models

We need to add the models of the Joanie entities in order to be able to
manipulate them

* ✨(frontend) add factories for models

In order to facilitate the mock of the application or requests, we create
factories

* ✨(frontend) add repositories for models

To generalize the way to request the API, we create repositories and an http
service that takes care of sending the right parameters for each request

* ✨(frontend) add hooks to request api

For consistency between projects, note richie, we use the useResources to
request the API

* 🤡(frontend) add many handlers to mock api requests

When running the tests or just to mock the application, we want to use requests
and simulate the return from the api with MSW and these handlers

* ✨(frontend) add presentational components

we add multiple components as a sortable list or select language input

* 💄(frontend) update DashboardLayout.tsx

Update the ui and DashboardLayout.tsx

* ✨(frontend) add product search input

In the course form, we can add a link to a product. So we need a search field.

* ✨(frontend) add translatable content component

In most forms, some fields are translatable. We must therefore modify the
accept-language of requests in real time in order to be able to save a
translation

* ✨(frontend) add certificates-definitions components

We need a list to view the certificate definitions as well
as its form to add or edit

* ✨(frontend) add courses components template

We need a list to view the template as well and its form to add or edit
course, and its useful search field for other forms

* ✨(frontend) add course-runs components template

We need a list to view the template as well and its form to add or edit
course-run

* ✨(frontend) add organizations components template

We need a list to view the template as well and its form to add or edit
organization, and its useful search field for other forms

* ✨(frontend) update the main next page

Several features are to be integrated, such as support for React-query, the
translation provider etc.

* ✨(frontend) add organization pages

Add organization management pages, i.e. the list, the creation and
edition page

* ✨(frontend) add course pages

Add course management pages, i.e. the list, the creation and
edition page

* ✨(frontend) add course-runs pages

Add course-runs management pages, i.e. the list, the creation and
edition page

* ✨(frontend) add certificate-definition pages

Add certificate-definition management pages, i.e. the list, the creation and
edition page

* ✨(frontend) add product page

add default product page

* 🚨(frontend) run and fix lint

The IC no longer passes because lint problems have been detected

* 💚(frontend) fix ci test-front-admin

In the IC, the tests of the fronts no longer pass.
It lacks environment variables and options for the tests to pass correctly.

* 🔧(backend) update CSRF_TRUSTED_ORIGINS

Now that dev mode can be directly in docker via port 8072,
we need to update this setting.

* 🌐(frontend) extract and compile translations

New translations have been added, so we run the extract
and compile commands