🔧(frontend) update next.config.js to not import all @mui modules (#300)
* 🔧(frontend) update next.config.js

In dev mode, a slowness problem has been detected. Each page reload was way too
long for the size of the app.

This was caused by importing all the icons from the @mui/material-icons package.

* 🔧(frontend) update .eslintrc.json

To avoid having a too heavy configuration and to decrease the DX, we prohibit
the top level import. We force the path-level import.