✨(core) use course product relation api endpoint instead of product
After some load test, it appears the old product endpoint was not efficient in
term of db accesses. Indeed, it does not scale well as the more it had target
courses, the more database queries increase.

In the same time, it appears easier to fetch course product relation instead of
product to improve db access and retrieve information. So we fully replaced the
old product api endpoint by an improved version of the course product relation
endpoint.