✨(models) allow generating certificate on products of type certificate
The method "get_or_generate_certificate" was written to products of type
"credential" in mind. The method should also work when the order concerns
a product of type certificate. The case is quite different because this
type of products do not have target courses and the certificate is in
this case emitted for the course related to the product.