♻️(models/api) allow defining number of seats via order groups
The maximum number of orders for a product feature did not allow to
sell a product in several rounds. Once a maximum defined, we had to
keep increasing it in order to continue selling the related product.

By introducing order groups, we can add a number of seats each time
we have a set of products to sell.