⚗️(api/admin) create custom SessionAuthentication class
The SessionAuthentication class provided by DRF does not return response with
`WWW-Authenticate` header to comply with the RFC. Indeed, there is no valid
value to provide to this header to instruct how to authenticate on a session
base authentication system.

As consequence, the HTTP response returned when user is not authenticated
is 403. For our use case, it's weird as this is complex to distinguish
unauthenticated requests form forbidden requests (no enough rights). As a
workaround, we decide to not respect the RFC on this point and provide an
arbitrary value 'Session' to the `WWW-Authenticate` header in order to get a 401
 response when user is not authenticated.