🐛(core) allow to submit a pending order
Currently, we are not able to submit a pending order that is weird. Indeed, an
order goes in pending state, when the payment has failed. So the user should be
able to resubmit this order to retry a payment. Furthermore, on submit, we
copy product target course relation into order, in order to prevent issues we
must exec this copy only when order is draft.