🐛(backend) change language cookie name for dev mode
When we work locally with richie and joanie, we have the same domain name.
When using richie, a "django_language" cookie is issued. However, this cookie
is also used when we work in the back office of Joanie while the BO has its own
language logic. So we change the name of the cookie to joanie in dev mode.