♻️(models) refactor save override and persist enrollment sync result
The enrollment state resulting from an LMS synchronization was not saved
in the "set" method and was left to the caller. This lead to repeated calls
to "save". While refactoring this, we realized that the override of the
save method in BaseModel had lead to unhealthy overrides in some models.
We prefer to abandon factorizing this override and do what has to be done
on each model.