✅(factories) fix course run factory when forcing state
We have recently introduced a "Params" field that holds a "ref_date".
This date is computed when the Python code first loads. When the full
test suite is run, the gap between the ref_date and "now" computed in
tests can be significant and lead to random test failures because a
course run created with a state ONGOING_OPEN can in fact be closed...