✨(models/api) add contract models and API endpoints
We want to be able to sign learning conventions with students.
On the model of what was done for certificates, we add a
"ContractDefinition" model for universities to define the content
of the contract they want students to sign and a "Contract" model
to record actual contracts for each student on an order.