✨(backend) add client signature provider for order
In order to sign a contract of an order where the product
has a contract definition, we've added an API client
with a signature provider. The signature provider offers us
actions to create a signature procedure, to upload a file,
start a process, delete a procedure and to get notified
once the document has been signed or refused by a student.
All the information about the signature procedure is saved
and gets updated overtime on the contract object.
Once the file is signed and completed, it sends the file to
the signer through a link in their email. The signature
provider that we will use in Joanie is called Lex Persona.
We have provided a base signature class, and a dummy class
as well. In order to interact with the API client, there
are three new routes : one for incoming webhook events,
one to see the contract, and finally one to submit for
signature an order.