🐛(models) fix auto enrollment with closed and specific course runs
When a course has several course runs, some of them being open and others
being closed, the query does not see that only one course run is open and
fails to enroll the user automatically.

The case of a product with target courses limited to specific course runs
is also not taken into account.

Fixes #389