🧑‍💻(backend) set django_get_or_create attribute in our factories
We are using Sequence to generate value for unique fields. As a drawback, it
generates errors when the Sequence context is reset but resources has been
previously persisted in database. To prevent it, we use the
`django_get_or_create` Meta attribute in our factories, in order to retrieve the
 existing resource in case of conflict