✨(api) allow filtering orders by state or product type exclusions
Up to now, if we want to see all orders except the ones that are
`canceled`, we had to include all other states in our querystring.
This is prone to future bugs if we later add a new state.
Since the intention is to exclude the `canceled` orders, our filters
should allow expressing it as close to the intention as possible,
which is an exclusion.
Same issue when filtering on product types.