🔧(Makefile) fix several commands
The help did not work as expected; indeed command names were not displayed.
Then the command `test` was not working anymore since we updated the
`test-back` command. Furthermore, we recently add a command to run admin
frontend tests. So the `test` command should run both `test-back` and
`admin-test` commands now.