✨(backend) add api endpoint to retrieve contract by ownership or access
A user must be able to retrieve contracts in two scenarios. As a learner, they
must be able to retrieve contract they own. Then as an organization
administrator / owner, they should be able to retrieve organization contracts.
In order to do that we add three endpoints, a simple one to retrieve contracts
by ownership, a second nested in organizations which allow to retrieve all
contracts relying on the given organization and a last one, also nested under
courses which allows to retrieve contracts relying on a course if the user
is administrator or owner of an organization which manages this course.

Co-authored-by: Samy Ait-Ouakli <samy.aitouakli@gmail.com>