✨(backend) generate zipfile of contracts pdf
University representatives need to retrieve the
contracts that are signed by the students. We have
added a new method on our signature backend classes
to fetch contracts that are signed in bytes PDF.
There is a new method that is responsible to generate
a zipfile from a collection of bytes PDFs.