🗃️(backend/core) link OrderGroup to CourseProductRelation
Currently an OrderGroup is linked to a Product. This has a main drawback; indeed
we think product to be generic and sold on course. So currently, if we apply an
OrderGroup to a Product and sold this product over many courses, the seat
restriction will be global that is weird. Instead we would like to apply
OrderGroup restrictions over a CourseProductRelation, this is the purpose of
this refactoring.