✨(backend) route to create and get zip archive
University representatives will require to get all the signed contracts
that are attached to an organization or from a course product relation.
We enable this action by providing two new routes allowing a user who
has access rights to an organization to fetch and generate a ZIP archive
of signed contracts through two new routes. The first route, creates
the demande in generating the ZIP archive of contracts if some are
signed. The second route allows a specific user with the correct access
rights to fetch the ZIP archive once it has been generated.