✨(backend) add synchronize_course_runs management command
In some case, we could need to manually trigger course run or product
synchronization. We are now able to do that through this management command.
Furthermore, in the hypothesis, later we would want to remove automatic
synchronization logic based on signals, we were able to trigger synchronization
by invoking this management command.