🔧(tray) use Staging django environment instead of Development
We recently installed a dev dependencies installed only in Development
environment. That's break the tray, as this one is using a joanie docker image
build with a production target. As dev dependencies are not installed, some job
fails with a ModuleNotFoundError. So in order to fix that, we decide to use
Staging django enviromnent as default value instead of Development.