💚(circleci) define properly django environment in job test-back
In the job test-back, we would like to use Test django environment instead of
the ContinuousIntegration one. We have overridden this value through the
cimg/python3.10 image environment key but it appears that this variable was
 overriden by the project environment variable defined into CircleCI project
settings. To override properly this env var, we have to use `environment` key of
 the job itself.

 This override silently failed as ContinuousIntegration django environment
 inherits from Test and does not override any value.