✨(backend) update auto enrollment logic according to contract state
Currently, we try to enroll user to single course runs opened for enrollment
when an order is validated. Now, we have introduce contract and we forbid
enrollment on not listed course runs until the student has signed its contract,
we have to udpate this logic. Indeed, if an order does not relies on a contract,
 we can keep the current logic but if a contract has to be signed we have to
 deferred this logic on contract signature.

Resolve #502