✨(backend) remote endpoint API on course run metrics
We want other services to be able to communicate with Joanie.
To secure the communication from server to server, you must use
a random generator of token and set it into Joanie's settings in
the configuration variable AUTHORIZED_API_KEYS.
The external server will use this token when it wants to request
our remote endpoint by applying the specific token into the
headers of the request. We have provided the first remote endpoint
API where we return as a dictionary the amount of enrollments once
a course run has reached its end date, and the amount of students
who decided to purchase the accessibility to unlock the certificate
of a free course run from a resource link as input.