✨(backend) add sign_contracts ability to Organization
A user is able to sign contracts for an organization according to its
organization access. That's why we add a new ability on the organization model
called `sign_contracts`. Then we refactor the get_abilities method of the
Contract model as we are now able to process it according to this new
Organization ability.