🧑‍💻(backend) improve dx of generate_jwt_tokens command
The command "generate_jwt_tokens" is really useful but, it could be improved to
ease the work of frontend developers on Richie. Indeed, for developer, richie
allows to use a fake authentication interface which just get value from an
object containing a keypair username/jwt_tokens. So we update the command to log
 in output this object, in this way, developer has just to copy/paste it to have
  fresh users.