🐛(back) add missing django dockerflow middleware
The middleware from django.dockerflow was missing. Without it, we no
health check endpoint and containers using it can not be started.
We choose to override this middleware to avoid the call to request.user
in process_response. Accessing to it created +1 sql query and also users
can be created thought the DelegatedJWTAuthentication authentication
backend and we don't want this. We want to control when the user is
really created.