♻️(backend) certificate replace with generic issuer method
For consistency in the code, we have decided to use the generic
method to `generate_document` that is located in
`joanie.core.utils.issuers`. The Certificate model now uses the
generic method to generate document in PDF bytes format instead
of using its model method `generate_document`.
Tests and lother ocations where the model method was called
were updated.