✨(backend) debug view template for Invoice
We have implemented a new class to handle the debug view
to preview our PDF document template for Invoice.
It inherits from our parent class `DebugTemplateView`. We've
provided a new route to preview the template and the developer
can parse a `pk` to preview the generated invoice. It provides by
default the preview of a credit note.