🧑‍💻(backend) add support of org signature to dummy backend
To ease developer experience, we have created a dummy signature backend in
charge to mimic the behavior of a real signature backend. We recently add
organization signature logic but we do not update logic of this backend so
it was not working when an organization tries to sign a contract.