✨(back) dedicated storage for easy_thumbnail using boto3
We want to use a S3 storage with easy_thumbnail. We are already using it
for the default storage, we just want to redefined the location.
Easy_thumbnail lib is still not using the storages settings introduced
in django 4 so we have to create a dedicated class extending boto3
class. Pylint is warning us to override some abstract method, but this
is not our job to do that, it should be handle by boto3.