🐛(backend) avoid create an order with an archived closed course run
While we've implemented the remote API for course run metrics,
we've found out it was possible to make an order while the course
run has ended. We do not want this behavior of creating an order with
a course that is in state `archived_closed`. To fix this issue, we have
added a condition into the `clean()` method of the model Order to verify
the `CourseState` of the `Course` that is attached to the `CourseRun` :
if the prioirity is equal or superior to the "archived' state, then
we should raise a Validation Error.