✨(backend) nested order viewset on courses to filter validated orders
A course member must be able to get the list of all the
users (learners) from a given course who have validated orders.
To add this feature, we've added a new nested order viewset into
the course's routes. This endpoint returns a list of orders
that has been serialized which shows a light version of
information about a course's order. We handle the query parameters
that are parsed in the URL of the request, which are :
'organization_id' and 'product_id' and 'course_product_relation_id'.
They will come in handy to filter when a product is present in
two courses for example, or if a course member only wants to get
the list of orders by the leaners that are attached to an organization.
Finally, he can also get the list of orders by parsing a course product
relation id.