🩹(backend) do not use user.last_name to create org signatory step
Our user does not have lastname fulfilled, indeed, on our authentication
platform, there is only a full name that we store into the first name field. To
submit a contract to organization signature, we were using the `user.last_name`
but as this field was blank, Lex Persona returned an error.