✨(backend) new properties for organization and address models
In the organization model, we needed new properties in
order to use them when generating the contract in PDF
format. We also added a new relationship between the
Organization model and the Address model. The Address model
now has a field 'organization'. We made the relation this
way to prevent the cases of when an organization has multiple
addresses. By applying the relation on the Address model,
we can benefit from the behaviour on user's address that has
been already implemented. Some model constraints were added
on the Address model, where an address can only be attached
to either a User or an Organization.