✨(backend) crud for address management for organizations
In order for a super admin to manage addresses for
organizations, we've added a new viewset with write
actions (create, update, partially update, delete).
We always verify if the relation exists aside from
the create action. In the create action, we verify
that the organization exists before proceeding to the
creation of the address object. We have also completed
the AdminOrganizationSerializer with the 'addresses'
attribute, which will allow a super admin to see all
the addresses that are attached to an organization in the
back office.