♻️(back) change default rest framework pagination class
The setting DEFAULT_PAGINATION_CLASS was using a pagination class
defined in rest_framework but we are using an other one in our viewsets.
To be coherent and to avoid to define it every time, we choose to change
the default class and remove the pagination property in evey viewset.