✨(backend) contract definition context update with new properties
Since we added new properties on the Address model and the
Organization model, we had to update the context when
generating the contract definition. All keys were added
into the construction of the context. There are also default
placeholders if there are no order for example to preview
the contract definition template.