✨(backend) new fields for User and Course models
To complete the context when generating a contract
definition, new data was required to be compliant
with the law. We added the user's phone number for
the case where we want the student to sign his document
with an SMS authentification to allow to sign the file.
We have added the effort field on the model Course, which
corresponds to the volume of hours required to follow a
specific course. The 'effort' for a course, is a
DurationField from Django.