🔧(celery) configure worker name to use ping command
The inspect ping command is a better solution for the liveness and
readyness than using the django check command. But to do that we have to
name every worker. To this, we weill use the hostname as uniqueness as
no other worker can run on the same host in our infrastructure.