🔧(celery) allow to configure default queue name
We are using a blue green infrastructure. And we have an issue with
celery in this blue green strategy. When a new stack is deploy, celery
will discover thew newly created worker and if a message is posted we
will share the same default queue name so the task can be execute
randomly on the current or next stack. By defining the default queue
name with the deployment stamp no queue can be shared anymore.