➕(backend) library `timedelta_isoformat` added to dependencies
Our Course model has a DurationField to represent the effort,
and we need to be able to work with python's `datetime.timedelta`
and ISO 8601 format of the effort value. This library will
allow us to translate from the `datetime.timedelta` to the
iso format, and from the iso format to `datetime.timedelta` type.
The advantage of using this library is because it's in pure-Python.