✨(backend) enhance admin API serializer for course serializers
Staff users will be able to manage the value of `effort` for a Course
object. They can now create/update a new course object with a value
that is a duration field. It also works when parsing an ISO 8601
formatted value, and it naturally works when parsing a `datetime.timedelta`
value.