♻️(backend) prevent iso8601_to_duration tag to fail
This template tag can be used with wrong values. Currently, when a wrong value
is passed an Error is raised that prevent the template to be rendered. In order
to fix that, we mimic the behaviour of date django template tag which return an
empty string when the provided value is not a valid value.