✨(backend) generate certificate API endpoint for an order
New admin backoffice API endpoint to trigger the generation
of a certificate for an order if the product type is
'credential' or 'certificate'. It only generates the
certificate if it does not yet exist. In other cases,
if it exists we do not generate the certificate,
and if it is not eligible, we don't do anything.